+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Upcoming
+
+### 0.15.0
 - Started enforcing unique system and component type names as it's needed to keep correct stats.  Anonymous types will get an automatically generated name.
 - Made component and system types available in `world.stats.components` and `world.stats.systems`, so you can easily get a list of all component and system types for debugging purposes (and to plug as input into other Becsy APIs).